backend/requirements.txt
# Kavach Backend Requirements

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# AI Engine
groq==0.4.0
openai>=1.0.0  # For NVIDIA NIM API compatibility

# PDF Generation
weasyprint==59.0

# DOCX Generation
python-docx==1.1.0

# Payment Processing
razorpay==1.4.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0